Use simple TestCase and TestCaseWithDatabase
For the mapper unit tests we only need laravel app bootstrapping no database
