+++ tests/Domain/Contacts/Repositories/EloquentContactRepositoryTestWithDatabase.php
@@ -10,9 +10,9 @@ use App\Domain\Contacts\Models\ContactModel;
 use App\Domain\Contacts\Repositories\EloquentContactRepository;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Ramsey\Uuid\Uuid;
-use Tests\TestCase;
+use Tests\TestCaseWithDatabase;
 
-final class EloquentContactRepositoryTest extends TestCase
+final class EloquentContactRepositoryTest extends TestCaseWithDatabase
 {
     private EloquentContactRepository $contactRepository;
