+++ tests/Domain/Subscriptions/Repositories/EloquentSubscriptionRepositoryTestWithDatabase.php
@@ -10,9 +10,9 @@ use App\Domain\Subscriptions\Repositories\EloquentSubscriptionRepository;
 use App\Domain\Subscriptions\Subscription;
 use App\Domain\Subscriptions\SubscriptionCategory;
 use Ramsey\Uuid\Uuid;
-use Tests\TestCase;
+use Tests\TestCaseWithDatabase;
 
-final class EloquentSubscriptionRepositoryTest extends TestCase
+final class EloquentSubscriptionRepositoryTest extends TestCaseWithDatabase
 {
     private EloquentSubscriptionRepository $subscriptionRepository;
