+++ tests/UiTiDv1/Jobs/ActivateConsumerListenerTestWithDatabase.php
@@ -16,11 +16,11 @@ use LogicException;
 use PHPUnit\Framework\MockObject\MockObject;
 use Psr\Log\NullLogger;
 use Ramsey\Uuid\Uuid;
-use Tests\TestCase;
+use Tests\TestCaseWithDatabase;
 use Tests\UiTiDv1\CreatesMockUiTiDv1ClusterSDK;
 use Tests\UiTiDv1\CreatesMockUiTiDv1Consumer;
 
-final class ActivateConsumerListenerTest extends TestCase
+final class ActivateConsumerListenerTest extends TestCaseWithDatabase
 {
     use CreatesMockUiTiDv1ClusterSDK;
     use CreatesMockUiTiDv1Consumer;
